+++ CMakeLists.txt
@@ -433,6 +433,7 @@ if(XPU_SPECIFIC_KERNELS_ENABLED)
       "csrc/xpu/lora/lora_shrink.cpp"
       "csrc/xpu/lora/lora_expand.cpp"
       "csrc/xpu/sycl/deepseek_scaling_rope.cpp"
+      "csrc/xpu/sycl/multimodal_rope.cpp"
       "csrc/xpu/grouped_gemm/grouped_gemm_interface.cpp"
       "csrc/xpu/utils.cpp"
       "csrc/xpu/gdn_attn/gdn_attn_interface.cpp")
