+++ csrc/xpu/ops.h
@@ -66,6 +66,15 @@ std::tuple<at::Tensor, at::Tensor> deepseek_scaling_rope(
     int64_t rotary_dim,
     bool is_neox);
 
+void multimodal_rotary_embedding(
+    torch::Tensor& positions,           // [num_mrope_sections, num_tokens]
+    torch::Tensor& query,
+    std::optional<torch::Tensor> key,
+    int64_t head_size,
+    torch::Tensor& cos_sin_cache,       // [max_position, rot_dim]
+    bool is_neox,
+    std::vector<int64_t> mrope_section); // host int list [num_mrope_sections]
+
 void gdn_attention(
     torch::Tensor& core_attn_out,
     torch::Tensor& z,
