+++ tests/register_ops.py
@@ -70,7 +70,7 @@ def multimodal_rotary_embedding(
     is_neox: bool,
     mrope_section: list[int],
 ) -> None:
-    torch.ops._C.multimodal_rotary_embedding(positions, query, key,
+    torch.ops._xpu_C.multimodal_rotary_embedding(positions, query, key,
                                              head_size, cos_sin_cache,
                                              is_neox, mrope_section)
